+++ tests/test_pickling.py
@@ -16,7 +16,7 @@ from .helpers import (
 
 def test_fitcontrolsbicop() -> None:
   original_controls = pv.FitControlsBicop()
-  original_controls.family_set = pv.itau
+  original_controls.family_set = pv.families.itau
   original_controls.parametric_method = "itau"
 
   # Serialize the object
@@ -44,7 +44,7 @@ def test_fitcontrolsbicop() -> None:
 def test_fitcontrolsvinecop() -> None:
   # Create an instance of FitControlsVinecop with some configuration
   original_controls = pv.FitControlsVinecop()
-  original_controls.family_set = pv.itau
+  original_controls.family_set = pv.families.itau
   original_controls.parametric_method = "itau"
 
   # Serialize the object
@@ -80,7 +80,7 @@ def test_fitcontrolsvinecop() -> None:
 
 
 def test_bicop() -> None:
-  original_bicop = pv.Bicop(pv.gaussian)
+  original_bicop = pv.Bicop(pv.families.gaussian)
   original_bicop.parameters = np.array([[0.5]])
 
   # Serialize the object
@@ -109,7 +109,7 @@ def test_rvinestructure() -> None:
 
 def test_kde1d() -> None:
   # Test with unfitted Kde1d object first
-  original_kde = pv.Kde1d(
+  original_kde = pv.utils.Kde1d(
     xmin=-5.0,
     xmax=5.0,
     type="continuous",
@@ -148,8 +148,8 @@ def test_vinecop() -> None:
   n = 1000
   u = pv.to_pseudo_obs(random_data(d, n))
 
-  controls = pv.FitControlsVinecop(family_set=[pv.gaussian])
-  assert controls.family_set == [pv.gaussian]
+  controls = pv.FitControlsVinecop(family_set=[pv.families.gaussian])
+  assert controls.family_set == [pv.families.gaussian]
   original_cop = pv.Vinecop.from_data(u, controls=controls)
 
   # Serialize the object