+++ tests/test_simulate.py
@@ -4,4 +4,4 @@ import pyvinecopulib as pv
 
 
 def test_simulate_uniform() -> None:
-  assert isinstance(pv.simulate_uniform(10, 2, False, [1, 2]), np.ndarray)
+  assert isinstance(pv.utils.simulate_uniform(10, 2, False, [1, 2]), np.ndarray)